Added a go multifactor authentication systemd autodeployer that can deploy applications automatically
With this script you can deploy code directly to production environment without having to put in credentials and running a lot of commands. It of course depends on if you set up sshd_config just like i have and uses the systemd. It also depends on if you use ubuntu. Its kinda hardcoded to one environment. I hope the script will come anyone in handy, for example with the PAM part.
Note there is not implemented known hosts and some things might need fix.